move findPurchaseInPurchaseHistory tests
